pyproject.toml
[project]
name = "tg-bookmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastlite>=0.2.3",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.36",
    "telethon>=1.42.0",
]
